requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
tqdm>=4.65.0
requests>=2.31.0 